refactor(#2302): extract full-index enumeration to data_access submodule
Move iterate_all_partitions_via_full_index into a new
data_access/full_index_scan.rs so it reuses that module's pub(super) helpers
(build_v5_parser, read_compressed_offset_window, sort_by_token_order) without
widening their visibility, and keeps partition_lookup.rs under the file-size
ratchet (campsite rule #1116). No behaviour change.
